requirements.txt
Flask
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Mail